requirements.txt
Jinja2==3.1.2
Markdown==3.4.4
MarkupSafe==2.1.3
PyYAML==6.0.1
watchdog==3.0.0